requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
sentence-transformers
faiss-cpu
streamlit
python-dotenv

